requirements.txt
requests==2.26.0
beautifulsoup4==4.9.3
pycryptodome==3.10.1
